% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pt_arrays.R
\name{create_pt_output}
\alias{create_pt_output}
\title{output}
\usage{
create_pt_output(db, pt_id)
}
\arguments{
\item{db}{pt_registry dataset}

\item{pt_id}{a single value from pt_ids identifiying a patient id}
}
\value{
array [pt, days, out] = [none, length(giorno_studio), 1]
}
\description{
dialy: array [pt, days, out] = [none, length(giorno_studio), 1]
}
\examples{
if (FALSE) create_pt_output(pt_registry, pt_ids[[1]])
}
