% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pt_arrays.R
\name{create_pt_ptnames}
\alias{create_pt_ptnames}
\title{ptnames}
\usage{
create_pt_ptnames(db, pt_id)
}
\arguments{
\item{db}{pt_names dataset}

\item{pt_id}{a single value from pt_ids identifiying a patient id}
}
\value{
array [pt, var] = [none, 2]
}
\description{
baseline: array [pt, var] = [none, 2]
}
\examples{
if (FALSE) create_pt_ptnames(pt_names, pt_ids[[1]])
}
