% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pt_arrays.R
\name{create_pt_trd}
\alias{create_pt_trd}
\title{trd}
\usage{
create_pt_trd(db, pt_id)
}
\arguments{
\item{db}{weaningsTRD dataset}

\item{pt_id}{a single value from pt_ids identifiying a patient id}
}
\value{
an array of dim
[pt, minutes, days, var] = [none, 1440, length(giorno_studio), ]
}
\description{
trd
}
\examples{
create_pt_trd(weaningsTRD, pt_ids)[[1]]
}
