% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_folders.R
\name{import_folder}
\alias{import_folder}
\title{Import weanings files in a folder}
\usage{
import_folder(.dir_path, what = c("TRD", "LOG"), verbose = FALSE)
}
\arguments{
\item{.dir_path}{(chr) path to the folder containing the TRDs files
to import.}

\item{what}{(chr, "TRD") code of the file to import, i.e., one of
"TRD" or "LOG."}

\item{verbose}{(lgl, FALSE) would you like to have additional
messages to be signaled?}
}
\value{
a \link[tibble:tibble-package]{tibble} with the imported signals
(i.e. the tabular content plus the date and the patient id) from
all the TRDs files inside the \code{.dir_path} folder.
}
\description{
Read and import all the information in weanings files in a provided
folder. It merges all the corresponding tables in a single one. It
adds to the merged table a column reporting the folder name imported
}
\note{
the function import TRD only, i.e., they do not need to be
the only ones inside the folder.
}
\examples{
\dontrun{
  library(weaning)

  trd_explicit <- file.path(get_data_path(), "AB") |>
  import_folder("TRD")

  trd_implicit <- file.path(get_data_path(), "AB") |>
   import_folder()  # "TRD" is the default

  identical(trd_implicit, trd_explicit)

  log_explicit_only <- file.path(get_data_path(), "AB") |>
  import_folder("LOG")

  trd_implicit
  log_explicit_only
}

}
