% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_folders.R
\name{import_folders}
\alias{import_folders}
\title{Import folders of weanings files folders}
\usage{
import_folders(
  .dir_path,
  what = "TRD",
  verbose = FALSE,
  patients2remove = character()
)
}
\arguments{
\item{.dir_path}{(chr) path to the folder containing the folders
which contain TRDs files
to import.}

\item{what}{(chr, "TRD") code of the file to import, i.e., one of
"TRD" or "LOG."}

\item{verbose}{(lgl, FALSE) would you like to have additional
messages to be signaled?}

\item{patients2remove}{(chr) vector of patient to not consider}
}
\value{
a \link[tibble:tibble-package]{tibble} with the imported signals
(i.e. the tabular content plus the date and the patient id) from
all the TRDs files inside the \code{.dir_path} folder.
}
\description{
Given a folder collecting weanings files grouped inside distinct
folders, it reads and imports all the information in the files
and it merges all the corresponding tables in a single one. It adds
to the merged table a column reporting the main folder name.
}
\examples{
\dontrun{
  library(weaning)

  import_folders(get_data_path(), "TRD")
  import_folders(get_data_path(), "LOG")
}
}
