% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_patients.R
\name{import_patients}
\alias{import_patients}
\title{Import Patient DB Excel file}
\usage{
import_patients(verbose = FALSE, testing_time = FALSE, test_path = "")
}
\arguments{
\item{verbose}{(lgl, FALSE) would you like to have additional
messages to be signaled?}
}
\value{
a \link[tibble:tibble-package]{tibble} with the imported data
(i.e. the tabular content with demographic and clinical info) from
the folder right outside the LOG and TRD folder.
}
\description{
Read and import in a table the clinical and demographic info
reported in an excel file of the weanings study.
}
\examples{
\dontrun{
  library(weaning)

  import_patients()
}
}
