% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_registry.R
\name{import_registry}
\alias{import_registry}
\title{Import Patient daily registry Excel file}
\usage{
import_registry(verbose = FALSE, testing_time = FALSE, test_path = "")
}
\arguments{
\item{verbose}{(lgl, FALSE) would you like to have additional
messages to be signaled?}
}
\value{
a \link[tibble:tibble-package]{tibble} with the imported data
(i.e. the tabular content with a row for each day for each patient)
from the folder right outside the LOG and TRD folder.
}
\description{
Read and import in a table daily info about clinical situation of
each patient in the weanings study, collected on an excel file.
}
\examples{
\dontrun{
  library(weaning)

  import_registry()
}
}
