% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_trd.R
\name{import_trd}
\alias{import_trd}
\title{Import TRD file}
\usage{
import_trd(.file_path, verbose = FALSE)
}
\arguments{
\item{.file_path}{(chr) path to the TRD file to import}

\item{verbose}{(lgl, FALSE) would you like to have additional
messages to be signaled?}
}
\value{
a \link[tibble:tibble-package]{tibble} with the imported signals
(i.e. the tabular content plus the date and the patient id) from
the \code{.file_path} TRD file.
}
\description{
Read and import in a table the signals reported in a TRD file from
weanings study. Among the signals, add a column reporting the
patient' ID as \code{id_pat} and the date of the recording as \code{date}.
}
