% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_trd.R
\name{import_trd_folder}
\alias{import_trd_folder}
\title{Import TRDs in a folder}
\usage{
import_trd_folder(.dir_path, verbose = FALSE)
}
\arguments{
\item{.dir_path}{(chr) path to the folder containing the TRDs files
to import.}

\item{verbose}{(lgl, FALSE) would you like to have additional
messages to be signaled?}
}
\description{
Read and import all the infromation in the TRDs fils in a provided
folder. It merge all the corresponding tables in a single one. It
adds to the merged table a column reporting the folder name imported
}
\note{
the function import TRD only, i.e., they do not need to be
the only ones inside the folder.
}
