% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trd.R
\name{plot_trd}
\alias{plot_trd}
\title{Build a plot of most important TRD variables}
\usage{
plot_trd(trd, weaning_subset = NA, color_files = FALSE, moving_avg = FALSE)
}
\arguments{
\item{weaning_subset}{(tibble) a tibble containing a list of SBT
with a "id_univoco", "data_lettura" and "esito" variable}

\item{color_files}{(logical) if TRUE, color the lines according to
the TRD file from which they originate. If FALSE, lines are colored
according to TRD variables}

\item{moving_avg}{(logical) if TRUE, a 30 minutes moving average
window is applied to TRD data}
}
\value{
a \link[ggplot2:ggplot2-package]{ggplot} showing the plot of
mechanical ventilation variables
}
\description{
Show on a single plot all relevant variables about a single day
for each patient to spot a weaning attempt:
}
\examples{
\dontrun{
  library(weaning)
  library(dplyr)

  patient_subset <- c("TS012", "BS002", "NO004")

  weaning_days <- tar_read(pt_registry) |>
    group_by( id_univoco) |>
    filter( susp_tot == 12,
            estubato == 0)  |>
    mutate( esito = factor( x = "Fallito",
                            levels = c("Fallito", "Successo")))

  weaning_subset <- weaning_days |>
    select(id_univoco, data_lettura, esito, type) |>
    filter( id_univoco \%in\% patient_subset) |>
    group_by(id_univoco) |>
    arrange(data_lettura)

  plot_trd( weaning_subset)

}

}
