% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/share_objects.R
\name{share_objects}
\alias{share_objects}
\title{Save object to a shared location}
\usage{
share_objects(obj_list, use_date_version = TRUE)
}
\arguments{
\item{obj_list}{(list) \code{{targets}} object(s) to share.}

\item{use_date_version}{(lgl, TRUE) do you want to prepend datetime
stamp (format: yyyymmddHHMM) to the file name
(separated by a dash, '-')?}
}
\value{
(chr) path(s) of the saved object(s)
}
\description{
Save all the objects listed in \code{obj_list} to the location defined
by the environmental variables \code{PRJ_SHARED_PATH} and
\code{OUTPUT_DATA_FOLDER}, i.e., to the
\code{file.path(Sys.getenv("PRJ_SHARED_PATH"), Sys.getenv("OUTPUT_DATA_FOLDER"))}.
}
