% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weaning-package.R
\docType{package}
\name{weaning-package}
\alias{weaning}
\alias{weaning-package}
\title{weaning: Weanings machine learning explorations at UBEP}
\description{
Collect data, functionalities and analyses for the weanings project at UBEP.
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Silvano Salaris \email{silvano.salaris@ubep.unipd.it} (\href{https://orcid.org/0000-0002-8389-7460}{ORCID})
}

Other contributors:
\itemize{
  \item Danila Azzolina \email{danila.azzolina@ubep.unipd.it} (\href{https://orcid.org/0000-0002-8185-5742}{ORCID}) [contributor]
  \item Giulia Lorenzoni \email{giulia.lorenzoni@ubep.unipd.it} (\href{https://orcid.org/0000-0003-1771-4686}{ORCID}) [contributor]
  \item Dario Gregori \email{dario.gregori@ubep.unipd.it} (\href{https://orcid.org/0000-0001-7906-0580}{ORCID}) [contributor, copyright holder]
  \item UBEP [copyright holder]
}

}
\keyword{internal}
