% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{weanings_log}
\alias{weanings_log}
\title{Weanings data LOG}
\format{
A data frame with 1.631.510 rows and 8 variables:
\describe{
\item{folder}{(chr) }
\item{file}{(chr) }
\item{id_pat}{(num) }
\item{data}{(Date) }
\item{ora}{{hms} }
\item{id_info}{(int) }
\item{tipo}{(fct) }
\item{informazioni}{(chr) }
}
}
\source{
...
}
\description{
A dataset containing the ....
}
