% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{weanings_trd}
\alias{weanings_trd}
\title{Weanings data}
\format{
A data frame with 2.483.787 rows and 33 variables:
\describe{
\item{folder}{(chr) main folder for the imported file, i.e., the center code}
\item{file}{ (chr) imported file name, pattern: CCDDD_XXXX where CC is the center code, DDD is teh patient id and XXXX is the sequence-code of the file.}
\item{id_pat}{(num) patient id as a number}
\item{date}{date of transmission}
\item{ora}{time of signals (granularity: 1 minute)}
\item{et_co2_percent}{(num) }
\item{eliminazione_co2_corrente_ml}{(num) }
\item{eliminazione_co2_minuto_ml_min}{(num) }
\item{caratteristiche_dinamiche_ml_cm_h2o}{(num) }
\item{elastanza_cm_h2o_l}{(num) a measure of the work that has to be exerted by the muscles of inspiration to expand the lungs [cmH2O/l]}
\item{flusso_di_fine_esp_l_min}{(num) }
\item{pressione_di_fine_esp_cm_h2o}{(num) }
\item{resistenza_esp_cm_h2o_l_s}{(num) }
\item{vol_minuto_espirato_l_min}{(num) }
\item{vol_corrente_espirato_ml}{(num) }
\item{concentraz_o2_percent}{(num) }
\item{resistenza_inspiratoria_cm_h2o_l_s}{(num) }
\item{vol_minuto_inspirato_l_min}{(num) }
\item{vol_corrente_inspirato_ml}{(num) }
\item{press_media_vie_aeree_cm_h2o}{(num) }
\item{frequenza_respiraz_misurata_resp_min}{(num) }
\item{freq_spontanea_resp_min}{(num) }
\item{picco_edi_m_v}{(num) }
\item{edi_min_m_v}{(num) }
\item{press_di_pausa_vie_aeree_cm_h2o}{(num) }
\item{press_di_picco_delle_vie_aeree_cm_h2o}{(num) }
\item{passa_a_backup_min}{(num) }
\item{backup_percent_min}{(num) }
\item{compliance_statica_ml_cm_h2o}{(num) }
\item{p_0_1_cm_h2o}{(num) }
\item{lavoro_respiratorio_del_ventilatore_joule_l}{(num) }
\item{lavoro_respiratorio_del_paziente_joule_l}{(num) }
\item{sbi}{(num) }
\item{volume_minuto_espirato_spontaneo_l_min}{(num) }
\item{perdita_percentuale_percent}{(num) }
\item{stress_index}{(num) }
\item{et_co2_mm_hg}{(num) }
}
}
\source{
...
}
\description{
A dataset containing the ....
}
