% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_trd.R
\name{import_trd_folders}
\alias{import_trd_folders}
\title{Import folders of TRDs folders}
\usage{
import_trd_folders(.dir_path, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Import folders of TRDs folders
}
